requirements.txt
requests==2.7.0
PyYAML==3.11